# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS «Медицина Долголетия» — API
  version: "1.0.0"
  description: |
    API платформы обучения врачей. Авторизация — через httpOnly cookie `token` (JWT),
    выдаётся эндпоинтами `/api/auth/login` и `/api/auth/register`.
    Роли: `student` (врач), `curator`, `admin`, `super_admin`.
servers:
  - url: /api
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /staff/team/{id}:
    delete:
      tags: [Staff]
      summary: Отозвать доступ у сотрудника (admin/super_admin, по иерархии ролей)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "403": { description: "Недостаточно прав" } }
  /staff/team/{id}/reset-password:
    post:
      tags: [Staff]
      summary: Сбросить пароль коллеге, вернёт временный пароль один раз
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { tempPassword: { type: string } }
...
  /staff/students/{id}:
    get:
      tags: [Staff]
      summary: Карточка врача
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "404": { description: "Не найдено" } }
  /staff/students/{id}/reset-password:
    post:
      tags: [Staff]
      summary: Сбросить пароль врачу, вернёт временный пароль один раз
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { tempPassword: { type: string } }
  /staff/students/{id}/access:
    patch:
      tags: [Staff]
      summary: Установить дату окончания доступа (null — снять срок)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { expiresAt: { type: string, format: date, nullable: true } }
      responses: { "200": { description: OK } }
  /staff/students/{id}/access/extend:
    post:
      tags: [Staff]
      summary: Продлить доступ на N дней
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { days: { type: integer } } }
      responses: { "200": { description: OK } }
  /staff/students/{id}/access/block:
    patch:
      tags: [Staff]
      summary: Заблокировать/разблокировать доступ немедленно
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { blocked: { type: boolean } } }
      responses: { "200": { description: OK } }
  /staff/students/{id}/stream:
    patch:
      tags: [Staff]
      summary: Назначить поток одному врачу
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /staff/students/bulk-field:
    post:
      tags: [Staff]
      summary: Массово проставить продукт или статус оплаты
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, items: { type: string } }
                field: { type: string, enum: [product, payment_status] }
                value: { type: string }
      responses: { "200": { description: OK } }
  /staff/audit-log:
    get:
      tags: [Staff]
      summary: Журнал действий персонала (admin/super_admin)
      parameters:
        - { name: limit, in: query, schema: { type: integer, maximum: 300 } }
      responses: { "200": { description: OK } }
...